requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
beautifulsoup4>=4.12.0
